Fix package name of "extensions" API
